+++ legacy_packages/chains/chains/47992.ts
@@ -1,0 +1,30 @@
+import type { Chain } from "../src/types";
+export default {
+  "chain": "Avalanche",
+  "chainId": 47992,
+  "explorers": [],
+  "faucets": [],
+  "features": [],
+  "icon": {
+    "url": "https://images.ctfassets.net/9bazykntljf6/62CceHSYsRS4D9fgDSkLRB/877cb8f26954e1743ff535fd7fdaf78f/avacloud-placeholder.svg",
+    "width": 256,
+    "height": 256,
+    "format": "svg"
+  },
+  "infoURL": "https://avacloud.io",
+  "name": "QaUser2555",
+  "nativeCurrency": {
+    "name": "QaUser2555 Token",
+    "symbol": "PRY",
+    "decimals": 18
+  },
+  "networkId": 47992,
+  "redFlags": [],
+  "rpc": [
+    "https://47992.rpc.thirdweb.com/${THIRDWEB_API_KEY}",
+    "https://subnets.avax.network/defi-kingdoms/dfk-chain/rpc"
+  ],
+  "shortName": "QaUser2555",
+  "slug": "qauser2555",
+  "testnet": true
+} as const satisfies Chain;