+++ packages/thirdweb/src/react/web/ui/ConnectWallet/Modal/ConnectModalContent.tsx
@@ -3,6 +3,7 @@ import type { Wallet } from "../../../../../wallets/interfaces/wallet.js";
 import { useSiweAuth } from "../../../../core/hooks/auth/useSiweAuth.js";
 import { useConnectUI } from "../../../../core/hooks/others/useWalletConnectionCtx.js";
 import { useConnect } from "../../../../core/hooks/wallets/wallet-hooks.js";
+import { useSetSelectionData } from "../../../providers/wallet-ui-states-provider.js";
 import { LoadingScreen } from "../../../wallets/shared/LoadingScreen.js";
 import { WalletSelector } from "../WalletSelector.js";
 import { onModalUnmount, reservedScreens } from "../constants.js";
@@ -39,6 +40,7 @@ export const ConnectModalContent = (props: {
     client,
   } = useConnectUI();
   const { connect } = useConnect();
+  const setSelectionData = useSetSelectionData();
 
   const siweAuth = useSiweAuth(auth);
   const showSignatureScreen = siweAuth.requiresAuth && !siweAuth.isLoggedIn;
@@ -52,6 +54,7 @@ export const ConnectModalContent = (props: {
       }
 
       onModalUnmount(() => {
+        setSelectionData({});
         setModalVisibility(true);
       });
 
@@ -69,6 +72,7 @@ export const ConnectModalContent = (props: {
       connect,
       showSignatureScreen,
       setScreen,
+      setSelectionData,
     ],
   );
 
@@ -89,6 +93,7 @@ export const ConnectModalContent = (props: {
       }}
       done={handleConnected}
       goBack={wallets.length > 1 ? handleBack : undefined}
+      setModalVisibility={setModalVisibility}
     />
   );