+++ legacy_packages/cli/CHANGELOG.md
@@ -1,5 +1,25 @@
 # thirdweb
 
+## 0.13.74
+
+### Patch Changes
+
+- Updated dependencies [[`45677ad`](https://github.com/thirdweb-dev/js/commit/45677ad62012c63e3443f2e236311eb9a2bad094)]:
+  - @thirdweb-dev/chains@0.1.99
+  - @thirdweb-dev/sdk@4.0.68
+  - @thirdweb-dev/wallets@2.5.8
+  - @thirdweb-dev/auth@4.1.66
+
+## 0.13.73
+
+### Patch Changes
+
+- Updated dependencies [[`6d6e0ac`](https://github.com/thirdweb-dev/js/commit/6d6e0ac69453d5a448ab771084976928313662ab)]:
+  - @thirdweb-dev/chains@0.1.98
+  - @thirdweb-dev/sdk@4.0.67
+  - @thirdweb-dev/wallets@2.5.7
+  - @thirdweb-dev/auth@4.1.65
+
 ## 0.13.72
 
 ### Patch Changes